procurement_bundle/pyproject.toml
[build-system]

requires = ["hatchling"]

build-backend = "hatchling.build"

[project]

name = "procurement"

version = "0.1.0"

description = "Procurement Lakeflow DLT Framework"

dependencies = [

  "databricks-labs-dlt"

]

[tool.hatch.build.targets.wheel]

packages = ["src/procurement"]
 